backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings>=2.0
python-dotenv
pytest
requests
psycopg2-binary
python-jose[cryptography]



